support url for image motion
